add some vehicle status messages (armed and rc_lost) to rc_channels module and fix issue with timestamp in matlab

+++ src/modules/udp_attitude/param.h
@@ -11,7 +11,7 @@ typedef struct{
   float yaw_s;
   float quat[4];
   float att_control[3];
-  unsigned int timestamp;
+  float timestamp;
 } attitudeValues;
 
 #endif
